change telescope and lualine
